Backend

+++ backend/src/member-data/products/products.service.ts
@@ -56,7 +56,7 @@ export class ProductsService {
       this.nullCheckForEntity(productEntity);
 
       const urlEntities:UrlEntity[] = await this.urlRepository.findBy({product: productEntity});
-      this.nullCheckForEntity(urlEntities)
+      if (urlEntities.length ==0) throw new NotFoundException()
 
       for(const urlEntity of urlEntities){
         await this.urlRepository.remove(urlEntity);
@@ -68,6 +68,13 @@ export class ProductsService {
         throw new BadRequestException({
           HttpStatus: HttpStatus.BAD_REQUEST,
           error: '[ERROR] 제품을 삭제하는 중 오류가 발생했습니다. id 형식이 올바른지 확인해주세요.',
+          message: '서버에 오류가 발생했습니다. 잠시후 다시 시도해주세요.',
+          cause: error,
+        });
+      } else if(error instanceof NotFoundException){
+        throw new NotFoundException({
+          HttpStatus: HttpStatus.NOT_FOUND,
+          error: '[ERROR] 제품을 삭제하는 중 오류가 발생했습니다. 해당 제품 및 url을 찾지 못했습니다.',
           message: '서버에 오류가 발생했습니다. 잠시후 다시 시도해주세요.',
           cause: error,
         });
